// lib/api/schema.ts
export interface components {
    schemas: {
// ...
        NotificationPublic: {
            /** Kind */
            kind: string;
            /** Body */
            body: string;
            /** Channel */
            channel: string;
            /** Status */
            status: string;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Ticket Id */
            ticket_id?: string | null;
            /** Position */
            position?: number | null;
            /** Created At */
            created_at?: string | null;
        };
        /** NotificationsPublic */
        NotificationsPublic: {
            /** Data */
            data: components["schemas"]["NotificationPublic"][];
            /** Count */
            count: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}